Splitting models into own package
