+++ modules/core/shared/src/main/scala/com/infixtrading/flashbot/core/FuzzyBook.scala
@@ -1,6 +1,8 @@
 package com.infixtrading.flashbot.core
 
 import com.infixtrading.flashbot.core.Order._
+import com.infixtrading.flashbot.models.core.{Order, OrderBook}
+
 import scala.concurrent.duration._
 
 /**
