+++ modules/core/shared/src/main/scala/com/infixtrading/flashbot/core/Instrument.scala
@@ -1,6 +1,7 @@
 package com.infixtrading.flashbot.core
 
 import com.infixtrading.flashbot.core.Order.{Buy, Fill, Sell}
+import com.infixtrading.flashbot.models.core.{Account, Portfolio}
 
 trait Instrument {
