+++ modules/core/shared/src/main/scala/com/infixtrading/flashbot/core/PriceIndex.scala
@@ -1,5 +1,7 @@
 package com.infixtrading.flashbot.core
 
+import com.infixtrading.flashbot.models.core.Market
+
 import scala.collection.immutable.{Map, Set}
 
 /**
