+++ modules/core/shared/src/main/scala/com/infixtrading/flashbot/core/package.scala
@@ -6,29 +6,6 @@ import java.time.Instant
 import scala.concurrent.duration.Duration
 
 package object core {
-  case class TimeRange(from: Long, to: Long = Long.MaxValue)
-  object TimeRange {
-    def build(now: Instant, from: String, to: String): TimeRange = {
-      val fromT = parseTime(now, from)
-      val toT = parseTime(now, to)
-      (fromT, toT) match {
-        case (Right(inst), Left(dur)) => TimeRange(
-          inst.toEpochMilli * 1000,
-          inst.plusMillis(dur.toMillis).toEpochMilli * 1000)
-        case (Left(dur), Right(inst)) => TimeRange(
-          inst.minusMillis(dur.toMillis).toEpochMilli * 1000,
-          inst.toEpochMilli * 1000)
-      }
-    }
-  }
-
-  def parseTime(now: Instant, str: String): Either[Duration, Instant] = {
-    if (str == "now") {
-      Right(now)
-    } else {
-      Left(parseDuration(str))
-    }
-  }
 
   sealed trait PairRole
   case object Base extends PairRole
