+++ modules/core/shared/src/main/scala/com/infixtrading/flashbot/models/core/Candle.scala
@@ -1,4 +1,5 @@
-package com.infixtrading.flashbot.core
+package com.infixtrading.flashbot.models.core
+import com.infixtrading.flashbot.core.Timestamped
 
 case class Candle(micros: Long,
                   open: Double,
