+++ modules/core/shared/src/main/scala/com/infixtrading/flashbot/models/core/Market.scala
@@ -1,4 +1,5 @@
-package com.infixtrading.flashbot.core
+package com.infixtrading.flashbot.models.core
+
 import io.circe.{KeyDecoder, KeyEncoder}
 
 case class Market(exchange: String, symbol: String) {
