+++ modules/core/shared/src/main/scala/com/infixtrading/flashbot/models/core/Position.scala
@@ -1,4 +1,6 @@
-package com.infixtrading.flashbot.core
+package com.infixtrading.flashbot.models.core
+
+import com.infixtrading.flashbot.core.Instrument
 import io.circe._
 import io.circe.generic.semiauto._
