+++ modules/server/src/main/resources/reference.conf
@@ -44,3 +44,35 @@ flashbot {
     configs = {}
   }
 }
+
+akka {
+  actor {
+    allow-java-serialization = off
+
+    serializers {
+      kryo = "com.twitter.chill.akka.AkkaSerializer"
+    }
+
+    serialization-bindings {
+      "io.circe.Json" = "com.infixtrading.flashbot.serializers.CirceAkkaSerializer"
+
+      "com.infixtrading.flashbot.models.core.Account" = kryo
+      "com.infixtrading.flashbot.models.core.Market" = kryo
+      "com.infixtrading.flashbot.models.core.Position" = kryo
+      "com.infixtrading.flashbot.models.core.FixedSize" = kryo
+      "com.infixtrading.flashbot.models.core.Portfolio" = kryo
+      "com.infixtrading.flashbot.report.Report" = kryo
+      "com.infixtrading.flashbot.report.ReportEvent" = kryo
+      "com.infixtrading.flashbot.report.ReportDelta" = kryo
+
+      "com.infixtrading.flashbot.models.core.TimeRange" = kryo
+      "com.infixtrading.flashbot.models.core.Candle" = kryo
+      "com.infixtrading.flashbot.models.core.Ladder" = kryo
+
+      "com.infixtrading.flashbot.models.api.TradingEngineState" = kryo
+      "com.infixtrading.flashbot.models.api.TradingSessionState" = kryo
+      "com.infixtrading.flashbot.models.api.TradingSessionEvent" = kryo
+      "com.infixtrading.flashbot.models.api.TradingSessionMode" = kryo
+    }
+  }
+}
