+++ modules/server/src/main/scala/com/infixtrading/flashbot/core/Exchange.scala
@@ -9,6 +9,7 @@ import io.circe.parser._
 import com.infixtrading.flashbot.core.Instrument.CurrencyPair
 import com.infixtrading.flashbot.core.Order.{Fill, Side}
 import com.infixtrading.flashbot.engine.TradingSession
+import com.infixtrading.flashbot.models.core.FixedSize
 
 import scala.collection.JavaConverters._
 import scala.math.BigDecimal.RoundingMode.HALF_DOWN
