+++ modules/server/src/main/scala/com/infixtrading/flashbot/core/TargetManager.scala
@@ -4,6 +4,8 @@ import com.infixtrading.flashbot.core.Action.{ActionQueue, CancelLimitOrder, Pos
 import com.infixtrading.flashbot.engine.IdManager
 import com.infixtrading.flashbot.engine.TradingSession.OrderTarget
 import com.infixtrading.flashbot.core.Convert._
+import com.infixtrading.flashbot.models.core.Action
+
 import scala.collection.immutable.Queue
 
 /**
