+++ modules/server/src/main/scala/com/infixtrading/flashbot/engine/DataSourceActor.scala
@@ -14,6 +14,7 @@ import com.infixtrading.flashbot.core.DataSource._
 import com.infixtrading.flashbot.core.MarketData.BaseMarketData
 import com.infixtrading.flashbot.core.Slice.SliceId
 import com.infixtrading.flashbot.core.FlashbotConfig._
+import com.infixtrading.flashbot.models.core.DataPath
 import com.infixtrading.flashbot.util.time._
 import com.infixtrading.flashbot.util.stream._
