+++ modules/server/src/main/scala/com/infixtrading/flashbot/engine/Strategy.scala
@@ -10,6 +10,7 @@ import com.infixtrading.flashbot.core.Convert._
 import com.infixtrading.flashbot.core.Instrument.CurrencyPair
 import com.infixtrading.flashbot.core._
 import com.infixtrading.flashbot.engine.TradingSession.{OrderTarget, SessionReportEvent}
+import com.infixtrading.flashbot.models.core._
 import com.infixtrading.flashbot.report.ReportEvent._
 
 import scala.concurrent.Future
