+++ modules/server/src/main/scala/com/infixtrading/flashbot/engine/TradingSession.scala
@@ -6,8 +6,8 @@ import io.circe.Json
 import com.infixtrading.flashbot.core.Action.ActionQueue
 import com.infixtrading.flashbot.core.DataSource._
 import com.infixtrading.flashbot.core._
+import com.infixtrading.flashbot.models.core.Portfolio
 import com.infixtrading.flashbot.report.{Report, ReportDelta, ReportEvent}
-
 
 trait TradingSession {
   def id: String
@@ -19,43 +19,6 @@ trait TradingSession {
 }
 
 object TradingSession {
-
-  trait Event
-  case class LogMessage(message: String) extends Event
-  case class OrderTarget(market: Market,
-                         key: String,
-                         size: FixedSize,
-                         price: Option[Double],
-                         once: Option[Boolean] = None,
-                         postOnly: Option[Boolean] = None) extends Event {
-    def id: String = s"$market:$key"
-  }
-  case class SessionReportEvent(event: ReportEvent) extends Event
-
-
-  sealed trait Mode
-  case class Backtest(range: TimeRange) extends Mode
-  case object Paper extends Mode
-  case object Live extends Mode
-
-  object Mode {
-    def apply(str: String): Mode = str match {
-      case "live" => Live
-      case "paper" => Paper
-    }
-  }
-
-  case class TradingSessionState(id: String,
-                                 strategy: String,
-                                 strategyParams: Json,
-                                 mode: Mode,
-                                 startedAt: Long,
-                                 portfolio: Portfolio,
-                                 report: Report) {
-    def updateReport(delta: ReportDelta): TradingSessionState =
-      copy(report = report.update(delta))
-  }
-
   case class SessionSetup(instruments: InstrumentIndex,
                           exchanges: Map[String, Exchange],
                           strategy: Strategy,
