+++ modules/server/src/main/scala/com/infixtrading/flashbot/exchanges/Simulator.scala
@@ -1,10 +1,10 @@
 package com.infixtrading.flashbot.exchanges
 
 import com.infixtrading.flashbot.core.Ladder.ladderFillOrder
-import com.infixtrading.flashbot.core.Order
 import com.infixtrading.flashbot.core.Order._
 import com.infixtrading.flashbot.core._
 import com.infixtrading.flashbot.engine.TradingSession
+import com.infixtrading.flashbot.models.core.{Ladder, Order, OrderBook}
 
 import scala.collection.immutable.Queue
