+++ modules/server/src/main/scala/com/infixtrading/flashbot/models/api/TradingSessionEvent.scala
@@ -1,0 +1,17 @@
+package com.infixtrading.flashbot.models.api
+import com.infixtrading.flashbot.models.core.{FixedSize, Market}
+import com.infixtrading.flashbot.report.ReportEvent
+
+trait TradingSessionEvent
+case class LogMessage(message: String) extends TradingSessionEvent
+case class OrderTarget(market: Market,
+                       key: String,
+                       size: FixedSize,
+                       price: Option[Double],
+                       once: Option[Boolean] = None,
+                       postOnly: Option[Boolean] = None) extends TradingSessionEvent {
+  def id: String = s"$market:$key"
+}
+case class SessionReportEvent(event: ReportEvent) extends TradingSessionEvent
+
+
