+++ modules/server/src/main/scala/com/infixtrading/flashbot/serializers/CirceAkkaSerializer.scala
@@ -1,0 +1,32 @@
+package com.infixtrading.flashbot.serializers
+import akka.serialization.Serializer
+import io.circe.Json
+import io.circe.parser._
+import io.circe.Printer
+
+class CirceAkkaSerializer extends Serializer {
+
+  // If you need logging here, introduce a constructor that takes an ExtendedActorSystem.
+  // class MyOwnSerializer(actorSystem: ExtendedActorSystem) extends Serializer
+  // Get a logger using:
+  // private val logger = Logging(actorSystem, this)
+
+  // This is whether "fromBinary" requires a "clazz" or not
+  def includeManifest: Boolean = false
+
+  // Pick a unique identifier for your Serializer,
+  // you've got a couple of billions to choose from,
+  // 0 - 40 is reserved by Akka itself
+  def identifier: Int = 100
+
+  // "toBinary" serializes the given object to an Array of Bytes
+  def toBinary(obj: AnyRef): Array[Byte] =
+    obj.asInstanceOf[Json].pretty(Printer.noSpaces).getBytes
+
+  // "fromBinary" deserializes the given array,
+  // using the type hint (if any, see "includeManifest" above)
+  def fromBinary(
+                  bytes: Array[Byte],
+                  clazz: Option[Class[_]]): AnyRef =
+    parse(new String(bytes)).right.get
+}
